% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_Index_computation_20240909.R
\name{axisPhylo_NL}
\alias{axisPhylo_NL}
\title{Equivalent of axisPhylo() in ape, with the option of relabeling the axis}
\usage{
axisPhylo_NL(
  side = 1,
  root.time = NULL,
  backward = TRUE,
  at_axis = NULL,
  lab_axis = NULL,
  ...
)
}
\description{
Equivalent of axisPhylo() in ape, with the option of relabeling the axis
}
