% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_Index_computation_20240909.R
\name{compute.timescale}
\alias{compute.timescale}
\title{Compute best timescale}
\usage{
compute.timescale(
  genome_length,
  mutation_rate,
  time_window,
  Ne = 1e+05,
  r_squared_threshold = 0.99
)
}
\arguments{
\item{genome_length}{Genome length (in bp)}

\item{mutation_rate}{Mutation rate (in mutation/site/year)}

\item{time_window}{Time window for gathering sequences in groups}

\item{Ne}{Effective population size}

\item{r_squared_threshold}{Threshold to find the best timescale}
}
\description{
Compute best timescale
}
