% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_Lineage_detection_20240909.R
\name{compute_fit_for_node_set}
\alias{compute_fit_for_node_set}
\title{Sub function to ouput a fit, for a selected number of nodes}
\usage{
compute_fit_for_node_set(
  timed_tree,
  metadata,
  node_set,
  weight_by_time = NULL,
  weighting_transformation = c(NULL, "inv_freq", "inv_sqrt", "inv_log"),
  k_smooth = 5,
  plot = T,
  log_y = T
)
}
\description{
Sub function to ouput a fit, for a selected number of nodes
}
