% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_Lineage_detection_20240909.R
\name{cumulative.index.all.nodes}
\alias{cumulative.index.all.nodes}
\title{Compute cumulative index of all nodes}
\usage{
cumulative.index.all.nodes(tree, index, time)
}
\arguments{
\item{tree}{Tree (object of class "phylo" or "multiPhylo")}

\item{index}{Index of each node}

\item{time}{Sampling (or inferred) time of each node}
}
\value{
List containing one vector per node, with the  descendant of each node
}
\description{
Compute cumulative index of all nodes
}
