% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_Lineage_detection_20240909.R
\name{descendants.all.nodes}
\alias{descendants.all.nodes}
\title{Output the descendants of all nodes (internal and external)}
\usage{
descendants.all.nodes(tree)
}
\arguments{
\item{tree}{Tree (object of class "phylo" or "multiPhylo")}
}
\value{
Vector containing the descendant of each node
}
\description{
Output the descendants of all nodes (internal and external)
}
