% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_Index_computation_20240909.R
\name{dist.nodes.with.names}
\alias{dist.nodes.with.names}
\title{Compute distance between all tips and nodes in the timed tree}
\usage{
dist.nodes.with.names(timed_tree)
}
\arguments{
\item{timed_tree}{Timed tree}
}
\value{
Distance matrix
}
\description{
Compute distance between all tips and nodes in the timed tree
}
