% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_Lineage_detection_20240909.R
\name{find.groups.by.index.dynamics}
\alias{find.groups.by.index.dynamics}
\title{Find groups defined by index dynamics}
\usage{
find.groups.by.index.dynamics(
  timed_tree,
  metadata,
  node_support = NULL,
  threshold_node_support = 0.5,
  time_window_initial = 1,
  time_window_increment = 0.5,
  min_descendants_per_tested_node = 100,
  min_group_size = 25,
  p_value_smooth = 0.01,
  stepwise_deviance_explained_threshold = 0.005,
  stepwise_AIC_threshold = 7,
  weight_by_time = NULL,
  weighting_transformation = c(NULL, "inv_freq", "inv_sqrt", "inv_log"),
  k_smooth = 5,
  plot_screening = F,
  parallelize_code = F,
  number_cores = 2,
  max_groups_found = 100,
  keep_track = F,
  log_y = T
)
}
\arguments{
\item{timed_tree}{Timed tree}

\item{metadata}{Metadata dataframe}

\item{node_support}{Numeric value of support for each node (e.g. mutations on the branch leading to the node, or bootstrap support)}

\item{threshold_node_support}{Threshold on the node support for the nodes to be considered in the detection algorithm}

\item{time_window_initial}{Time in years of when to start the screen}

\item{time_window_increment}{Window}

\item{min_descendants_per_tested_node}{To start the analysis, start from nodes that have this minimum number of sequences}

\item{min_group_size}{Minimum group size, when creating a new potential split}

\item{p_value_smooth}{p-value threshold to consider a new spline as significant.}

\item{stepwise_deviance_explained_threshold}{Change in explained deviance, recommend 0.005 (0.5 percent): By how much the new split should increase the percentage of deviance explained, at the very least.}

\item{stepwise_AIC_threshold}{AIC threshold, commonly 7: By how much the new split should increase the goodness of fit, at the very least.}

\item{weight_by_time}{NULL or numeric (in years): Size of the window of time on which to compute the weights}

\item{weighting_transformation}{NULL, inv_freq, inv_sqrt, or inv_log: Type of weighting to use.}

\item{k_smooth}{Number of knots for the splines}

\item{plot_screening}{TRUE or FALSE, to plot the discovered dynamics as soon as they're found}

\item{parallelize_code}{TRUE or FALSE, to parallelise code}

\item{number_cores}{Integer, if parallelised, how many cores to use}

\item{max_groups_found}{Integer, Maximum number of groups to find}

\item{keep_track}{TRUE or FALSE, whether user wants to output all the steps}

\item{log_y}{TRUE or FALSE, whether user wants to fit the index values on a log scale (TRUE), or with a log link (FALSE). The algorithm with the log link is much slower.}
}
\value{
Potential split nodes at target, and if keep_track==T, all the different nodes at each iteration
}
\description{
This code is fitting a generalised additive model (gam) to the index dynamics, trying to find the smallest number of lineages to explain best the index dynamics.
Model targets are AIC, explained deviance or maximum number of groups
The script is run until a target is met, and the user can choose to record all the different states (keep_track), or not
}
\details{
Potential nodes can be filtered out by support and threshold_node_support. This can be a minimum number of mutations on the branch leading to the node, or a bootstrap support.
}
