% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_Index_computation_20240909.R
\name{index.bandwidth}
\alias{index.bandwidth}
\title{Compute the index bandwidth from a set timescale, genome length and mutation rate}
\usage{
index.bandwidth(timescale, genome_length, mutation_rate)
}
\arguments{
\item{timescale}{Index timescale (in years)}

\item{genome_length}{Genome length (in bp)}

\item{mutation_rate}{Mutation rate (in mutation/site/year)}
}
\value{
Index bandwidth
}
\description{
Compute the index bandwidth from a set timescale, genome length and mutation rate
}
\references{
Wirth, T, Wong, V, Vandenesch, F, Rasigade, J-P. Applied phyloepidemiology: Detecting drivers of pathogen transmission from genomic signatures using density measures. Evol Appl. 2020; 13: 1513– 1525. https://doi.org/10.1111/eva.12991
}
