% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_Index_computation_20240909.R
\name{matrices.computation}
\alias{matrices.computation}
\title{Computes a matrix of pairwise distances for all pairs of isolates sampled at within the same window of time}
\usage{
matrices.computation(
  time_distance_mat,
  timed_tree,
  time_window,
  metadata,
  mutation_rate,
  genome_length
)
}
\arguments{
\item{time_distance_mat}{Hamming distance matrix}

\item{timed_tree}{Timed tree}

\item{time_window}{Time window for the index computation}

\item{metadata}{Metadata for all tips and nodes (see details)}

\item{mutation_rate}{Mutation rate (in mutation/site/year)}

\item{genome_length}{Genome length (in bp)}
}
\description{
Computes a matrix of pairwise distances for all pairs of isolates sampled at within the same window of time
}
