% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_Lineage_detection_20240909.R
\name{number.descendants.all.nodes}
\alias{number.descendants.all.nodes}
\title{Compute number of descendants of all nodes (internal and external)}
\usage{
number.descendants.all.nodes(tree)
}
\arguments{
\item{tree}{Tree (object of class "phylo" or "multiPhylo")}
}
\value{
Vector containing the number of descendants of each node
}
\description{
Compute number of descendants of all nodes (internal and external)
}
