% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_Lineage_detection_20240909.R
\name{path.root.to.all.nodes}
\alias{path.root.to.all.nodes}
\title{Compute the path from the tree root to all nodes}
\usage{
path.root.to.all.nodes(tree)
}
\arguments{
\item{tree}{Tree (object of class "phylo" or "multiPhylo").}
}
\value{
List with each element being a vector containing the path fromß the root to the node.
}
\description{
Compute the path from the tree root to all nodes
}
