% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_Lineage_detection_20240909.R
\name{run_gam_model}
\alias{run_gam_model}
\title{Sub function to run the generalised additive model (gam)}
\usage{
run_gam_model(
  metadata_sub,
  clade_data,
  weights_sub,
  k_smooth,
  coefs,
  sp,
  log_y = T
)
}
\description{
Sub function to run the generalised additive model (gam)
}
