% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_Lineage_detection_20240909.R
\name{test_node_gam}
\alias{test_node_gam}
\title{Sub function to test a node}
\usage{
test_node_gam(
  i,
  metadata_sub,
  metadata_sub_rownames,
  nodes_to_test_tmp,
  descendants_all_nodes,
  min_group_size,
  clade_membership_recorded,
  clade_data_null,
  weights_sub,
  k_smooth,
  mod_null,
  log_y = T
)
}
\description{
Sub function to test a node
}
