% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_Index_computation_20240909.R
\name{theoretical_index}
\alias{theoretical_index}
\title{Compute the theoretical index for a population of constant size Ne during times}
\usage{
theoretical_index(bandwidth, genome_length, mutation_rate, Ne, times)
}
\arguments{
\item{bandwidth}{Index timescale (in years)}

\item{genome_length}{Genome length (in bp)}

\item{mutation_rate}{Mutation rate (in mutation/site/year)}

\item{Ne}{Effective population size}

\item{time}{Time at which to compute the theoretical index (can be a double or vector)}
}
\description{
Compute the theoretical index for a population of constant size Ne during times
}
